rename file [Server] to [server], rename file [HttpSession] to [http_session], and fix compile error
