+++ src/http/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_library(http STATIC "")
 target_sources(http PRIVATE
-        ${CMAKE_CURRENT_SOURCE_DIR}/HttpSession.cc
-        ${CMAKE_CURRENT_SOURCE_DIR}/HttpSession.h)
+        ${CMAKE_CURRENT_SOURCE_DIR}/http_session.cc
+        ${CMAKE_CURRENT_SOURCE_DIR}/http_session.h)
 target_include_directories(http PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 target_link_libraries(http base)
