+++ src/http/http_session.cc
@@ -10,7 +10,7 @@
 #include <cerrno>
 #include <cstring>
 
-#include "HttpSession.h"
+#include "http_session.h"
 #include "utils.h"
