+++ src/main.cc
@@ -1,4 +1,4 @@
-#include "Server.h"
+#include "server.h"
 
 
 int main()
